Sphinx-based documentation
Adds a Sphinx-based documentation and enhances the existing documentation with a diagram of the CUDS objects that are used as input of the simulation and produced as output. Several sections from the README have been moved to the documentation. Text has been revised and adapted to the current status of the wrapper.